requirements.txt
aiogram==3.2
ollama
wikipedia-api
asyncpg
python-dotenv
